vcsa: first round of unwinding this role
This role was using an interesting method of dynamic play
generation, but Ansible has removed the ability to include
files based on variables. That's a long story, but the change
effectively requires revamping things a bit so as to include
explicit play sets -- in effect, all hosts must receive all
of the same plays, so generated include file paths no longer
support that in all cases.

Change-Id: I104ee1bf9c55c71a08e7f31023391b3abba8222a